(refactor) moved index.js to app.js, index.js now entry point for app / mysql db
